+++ packages/profile/src/components/context/index.tsx
@@ -1,2 +1,1 @@
-export { ThemeContext } from "./theme";
 export { Provider } from "./provider";
