+++ packages/profile/src/hooks/context.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { ThemeContext } from "@/components/context";
+import { ThemeContext } from "@/context/theme";
 import { ConnectionContext } from "@/context/connection";
 import { DataContext } from "@/context/data";
